+++ test/samovar/flags.rb
@@ -26,4 +26,20 @@ describe Samovar::BooleanFlag do
 		expect(flag.prefix?("--no-color")).to be == true
 		expect(flag.prefix?("--other")).to be == false
 	end
+	
+	it "ignores tokens that do not match the prefix" do
+		input = ["--other=true"]
+		expect(flag.parse(input)).to be_nil
+		expect(input).to be == ["--other=true"]
+	end
+end
+
+describe Samovar::ValueFlag do
+	let(:flag) {Samovar::Flag.parse("--config <path>")}
+	
+	it "ignores tokens that do not match the prefix" do
+		input = ["--other=value"]
+		expect(flag.parse(input)).to be_nil
+		expect(input).to be == ["--other=value"]
+	end
 end